Add files via upload
Here is a partially completed version of the program, I am at the fitness function stage of writing it. I have updated runForge.bat to run a Swiss style tournament and output the match results to the forge.log file. I think I may have to have my program freeze after runForge.bat because it does not wait for the forge.log file to be completed, and Forge runs simultaneously with my program.